Added config.py
config.py includes all important paths.
Read-only paths are relative paths to the main scripts
Paths to write to are determmined by getting the system specific AppData folder and creating a VolumanXR folder. Important when scripts are bundle with pyinstaller at a later stage as the temp folder the bundle is undbled in is read-only and gets deleted when the program ends